added playlist controllers and routes

+++ src/controllers/comment.controller.js
@@ -2,7 +2,7 @@ import mongoose, { Schema } from "mongoose";
 import { Comment } from "../models/comment.model.js";
 import { Video } from "../models/video.model.js";
 import { Like } from "../models/like.model.js";
-import ApiError from "../utils/apiError.js";
+import ApiError from "../utils/ApiError.js";
 import ApiResponse from "../utils/ApiResponse.js";
 import asyncHandler from "../utils/asyncHandler.js";
 
@@ -161,7 +161,11 @@ const deleteComment = asyncHandler(async(req, res) => {
         throw new ApiError(403, "You are not authorized to delete this comment")
     }
 
-    await Comment.findByIdAndDelete(commentId)
+    const isDeleted = await Comment.findByIdAndDelete(commentId)
+
+    if(!isDeleted){
+        throw new ApiError(500, "Failed to delete comment please try again");
+    }
 
     await Like.deleteMany({
         comment: commentId
